requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.8.0
dj-database-url==2.3.0
Django==4.2.16
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-environ==0.9.0
django-recaptcha==4.0.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
gunicorn==23.0.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
pytest

